chore: rename module to builder.rs
