+++ bridge-fungible-token/Forc.lock
@@ -19,10 +19,10 @@ dependencies = ['std']
 
 [[package]]
 name = 'core'
-source = 'path+from-root-2F6DBEF6F2B7B5D1'
+source = 'path+from-root-7769F7A51C4F10C0'
 dependencies = []
 
 [[package]]
 name = 'std'
-source = 'git+https://github.com/fuellabs/sway?tag=v0.26.0#e7674f704f2706e22f77c0ed32df9c89302e5e7e'
+source = 'git+https://github.com/fuellabs/sway?tag=v0.27.0#40f1e79de08af109f2ebb066f51704f36094ed1e'
 dependencies = ['core']
