+++ bridge-fungible-token/tests/utils/environment.rs
@@ -1,4 +1,4 @@
-use crate::ext_sdk_provider;
+use crate::builder;
 
 use std::mem::size_of;
 use std::num::ParseIntError;
@@ -71,8 +71,7 @@ pub async fn setup_environment(
         Some(v) => Address::from_str(v).unwrap(),
         None => Address::from_str(MESSAGE_SENDER_ADDRESS).unwrap(),
     };
-    let (predicate_bytecode, predicate_root, _) =
-        ext_sdk_provider::get_contract_message_predicate().await;
+    let (predicate_bytecode, predicate_root, _) = builder::get_contract_message_predicate().await;
     let all_messages: Vec<Message> = messages
         .iter()
         .flat_map(|message| {
@@ -172,7 +171,7 @@ pub async fn relay_message_to_contract(
     optional_outputs: &[Output],
 ) -> Vec<Receipt> {
     // Build transaction
-    let mut tx = ext_sdk_provider::build_contract_message_tx(
+    let mut tx = builder::build_contract_message_tx(
         message,
         contract,
         gas_coins,
